requirements.txt
fastapi
uvicorn
pinecone-client
openai
python-dotenv
